+++ packages/app/index.js
@@ -5,6 +5,7 @@ import '@formatjs/intl-pluralrules/locale-data/es.js';
 import '@formatjs/intl-pluralrules/locale-data/fr.js';
 import '@formatjs/intl-pluralrules/locale-data/uk.js';
 import './src/account/task/account-balance-incremental.task';
+import './src/budget/task/budget-alert-monitor.task';
 import './src/exchange-rate/task/exchange-rate-sync.task';
 import './src/sync/task/monobank-sync.task';
 import './src/sync/task/binance-sync.task';